+++ restygwt/src/main/java/org/fusesource/restygwt/client/callback/FilterawareRetryingCallback.java
@@ -34,11 +34,11 @@ import com.google.gwt.user.client.Window;
 
 public class FilterawareRetryingCallback implements FilterawareRequestCallback {
 
+    public static int DEFAULT_NUMBER_OF_RETRIES = 5;
     /**
      * Used by RetryingCallback
-     * default value is 5
      */
-    protected static int numberOfRetries = 5;
+    protected int numberOfRetries;
 
     /**
      * time to wait for reconnect upon failure
@@ -54,10 +54,16 @@ public class FilterawareRetryingCallback implements FilterawareRequestCallback {
     final protected List<CallbackFilter> callbackFilters = new ArrayList<CallbackFilter>();
 
     public FilterawareRetryingCallback(Method method) {
+        // default number of retries is 5
+        this(method, DEFAULT_NUMBER_OF_RETRIES);
+    }
+
+    public FilterawareRetryingCallback(Method method, int numberOfRetries) {
         this.method = method;
         // need to keep requestcallback here, as ``method.builder.getCallback()`` does not
         // give the same callback later on
         this.requestCallback = method.builder.getCallback();
+        this.numberOfRetries = numberOfRetries;
     }
 
     @Override
@@ -89,9 +95,7 @@ public class FilterawareRetryingCallback implements FilterawareRequestCallback {
                  *  RuntimeException token from
                  *  com.google.gwt.http.client.Request#fireOnResponseReceived()
                  */
-                requestCallback.onError(request, new RuntimeException("Response "
-                        + response.getStatusCode() + " for " + method.builder.getHTTPMethod() + " "
-                        + method.builder.getUrl()));
+                requestCallback.onResponseReceived(request, response);
             }
             return;
         } else {
@@ -165,10 +169,15 @@ public class FilterawareRetryingCallback implements FilterawareRequestCallback {
             if (null != request
                     && null != response
                     && null != requestCallback) {
+
+                if (LogConfiguration.loggingIsEnabled()) {
+                    Logger.getLogger(FilterawareRetryingCallback.class.getName()).severe("Response "
+                          + response.getStatusCode() + " for " + method.builder.getHTTPMethod() + " "
+                          + method.builder.getUrl() + " after " + numberOfRetries + " retries.");
+                }
+
                 // got the original callback, call error here
-                requestCallback.onError(request, new RuntimeException("Response "
-                        + response.getStatusCode() + " for " + method.builder.getHTTPMethod() + " "
-                        + method.builder.getUrl() + " after " + numberOfRetries + " retries."));
+                requestCallback.onResponseReceived(request, response);
             } else {
                 // got no callback - well, goodbye
                 if (Window.confirm("error")) {